llama : move grammar code into llama-grammar
ggml-ci
