+++ src/CMakeLists.txt
@@ -14,6 +14,8 @@ endif()
 add_library(llama
             ../include/llama.h
             llama.cpp
+            llama-vocab.cpp
+            llama-grammar.cpp
             llama-sampling.cpp
             unicode.h
             unicode.cpp
